tests: import sedparse only once
Remove my ugly hack that was used to import sedparse again before
every single test. Now it is imported only once at the top.

Gone are the methods _my_setup() and _my_tear_down(), and the tests
for the error messages needed to be adapted because before they relied
on the fact that every tested script was always the first expression.