Add __init__() to all structs
It was a mistake not adding that from the start. Now leaking of data
between commands is not a worry anymore.

The test test_command_data_cleanup() is probably void now, but I'll
leave it to make sure any regression will not happen on that matter.

The previous code to reset all the structs in next_cmd_entry() was
removed, because it is not necessary anymore.