Now dumps a JSON for the parsed data
This is more useful to those needing to parse a sed script, and is
language and platform agnostic.

This is now the default action when informing a sed script file in
the command line, and can also be called directly as
`sedparse.to_json()`, informing the object to serialize.

Note that the old sample.out file was now divided into two, one JSON
file and one file only with the debug messages.