Refactor sample.sed to reference.sed
All the tricky tests in the old sample.sed file are now in the proper
unit tests of tests/test_parser.py.

But there's a need for a general reference file, where users can
quickly see how a specific sed command will be mapped to JSON (or the
internal structures, which are the same).

So the sed file now only contains commands that show different aspects
of the parsed data. Also, instead of only showing one command of a
kind, all the supported sed commands are listed there. For example,
only 'b' would be enough to understand how 't' and 'T' will be parsed,
but listing every command is easier as a reference.