Fix detecting } at the end of a label
The } ended the label, but it was not saved and detected as the next
command. So invalid scripts such as 'b}' were not considered syntax
error, because the '}' was not considered a command.

Now it's fixed, with the code in `read_label()` being more true to the
original: `savchar()` is always called.

New tests were added to `test_errors.py` to make sure the syntax error
is catched.

Tests for labels were also improved to not contain invalid scripts
anymore and handle the special case when the label is not the first
command in the script.