Add setup.py and other package-related changes
New `setup.py` and `setup.cfg` to generate the package to be used in
pypi.

In `sedparse.py` a new `entrypoint()` function was created to be the
entrypoint for the `sedparse` executable generated by `setup.py`.

New `requirements-pkg.txt` file to list the dependencies necessary to
build the sedparse package.

Contributor information was moved from the README.md file to a new
CONTRIBUTING.md file.